PCHR-1133:  Move getWorkTypeForDate to ContactWorkPattern BAO, Clean up LeaveRequest BAO
